refactor(stream): remove createRetryableStreamText, fold tests into createRetryableStream
Drop the streamText drop-in wrapper and its dedicated export; the generic
createRetryableStream is now the single entry point. The streamText
integration coverage (error-based retries, deadlines, deferred consumption,
callbacks, model-layer contrast) moves into create-retryable-stream.test.ts
behind an inline streamText glue.
