--- src/experimental/stream-text/index.ts
@@ -1,5 +1,0 @@
-export {
-  createRetryableStreamText,
-  type RetryableStreamText,
-  type RetryableStreamTextOptions,
-} from './create-retryable-stream-text.js';
